pyproject.toml
[tool.poetry]
name = "decopon"
version = "0.1.0"
description = ""
authors = ["matt <matt76k@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pygame = "^2.5.2"
pymunk = "^6.6.0"
gym = "^0.26.2"
pillow = "^10.1.0"
torch = "^2.1.1"
tqdm = "^4.66.1"
matplotlib = "^3.8.2"
torchvision = "^0.16.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.7.0"
isort = "^5.12.0"
pyproject-flake8 = "^6.0.0.post1"
mypy = "^1.4.1"
poethepoet = "^0.22.0"

[tool.flake8]
extend-ignore = "E203,E501,"

[tool.black]
line-length = 120
target-version = ['py310']

[tool.isort]
py_version = 310
line_length = 120
profile = "black"

[tool.mypy]
ignore_missing_imports = true

[tool.poe.env]
SRC_DIR = "src decopon"

[tool.poe.tasks]
lint = "poetry run pflake8 $SRC_DIR"
black = "poetry run black $SRC_DIR"
isort = "poetry run isort $SRC_DIR"
format = ["isort", "black", "lint"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"